Replacing quinn with quiche with streams

+++ client/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod client;
 
-pub const DEFAULT_MAX_STREAM: u32 = quic_geyser_common::quic::configure_client::DEFAULT_MAX_STREAMS;
+pub const DEFAULT_MAX_STREAM: u64 = quic_geyser_common::quic::configure_client::DEFAULT_MAX_STREAMS;
